// src/cast/types.ts
export interface CastDevice {
  /** Friendly device name (e.g., "Living Room Speaker") */
  name: string;
  /** IP address of the device */
  host: string;
  /** Cast port (typically 8009) */
  port: number;
  /** Unique device identifier from mDNS TXT record */
  id?: string;
}

/**
 * Cast metadata types
 * @see https://developers.google.com/cast/docs/reference/web_receiver/cast.framework.messages.MetadataType
 */
export enum MetadataType {
  /** Generic media content */
  GENERIC = 0,
  /** Movie */
  MOVIE = 1,
  /** TV Show */
  TV_SHOW = 2,
  /** Music Track */
  MUSIC_TRACK = 3,
  /** Audiobook Chapter - enables Nest Hub display-off in low light */
  AUDIOBOOK_CHAPTER = 4,
  /** Photo */
  PHOTO = 5,
}